requirements.txt
requests
python-dotenv
rich
transformers
torch
accelerate